Move stuff into CutUtils (#45)
Move various things out of Cuts.* and into CutUtils. Include Cuts vector, GetCutName, get various sub-lists of cuts. And GetHadIdxsFromMichels.

* Move stuff into CutUtils.

Just want cut definitions and cutting mechanics in the Cuts.* files.

* Revert for cleaner PR.

* Fix cut vector mistakes. Move CutUtils to .h.
